feat(app) Fix profile creation feat(app) Fix dates in AGs feat(demo) Fix iframe permissions
