+++ .changeset/green-clubs-report.md
@@ -1,0 +1,5 @@
+---
+"pay-demo": patch
+---
+
+Fix iframe permissions
