+++ .changeset/itchy-oranges-trade.md
@@ -1,0 +1,5 @@
+---
+"data-dashboard": patch
+---
+
+Small fixes with creation account or dates in AGs
