+++ apps/idos-data-dashboard/package.json
@@ -40,7 +40,7 @@
     "@tanstack/react-table": "catalog:tanstack",
     "@vercel/react-router": "catalog:react-router",
     "@wagmi/core": "catalog:ethereum",
-    "@xstate/react": "^6.0.0",
+    "@xstate/react": "^6.1.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "ethers": "catalog:ethereum",
@@ -67,7 +67,7 @@
     "zod": "catalog:core"
   },
   "devDependencies": {
-    "@near-js/types": "^0.3.0",
+    "@near-js/types": "^0.3.1",
     "@playwright/test": "^1.60.0",
     "@react-router/dev": "catalog:react-router",
     "@tailwindcss/vite": "^4.3.0",
@@ -77,7 +77,7 @@
     "@testing-library/react": "^14.3.1",
     "@types/jsonwebtoken": "^9.0.10",
     "@types/node": "catalog:build",
-    "@types/pg": "^8.11.0",
+    "@types/pg": "^8.20.0",
     "@types/react": "catalog:react",
     "@types/react-dom": "catalog:react",
     "dotenv": "^17.4.2",
@@ -87,7 +87,7 @@
     "tw-animate-css": "^1.4.0",
     "typescript": "catalog:build",
     "vite": "catalog:build",
-    "vite-plugin-mkcert": "^1.17.5",
+    "vite-plugin-mkcert": "^1.17.12",
     "vite-plugin-node-polyfills": "catalog:build",
     "vitest": "catalog:testing"
   }
