+++ apps/idos-data-dashboard/src/components/credentials/grants-center.tsx
@@ -11,7 +11,7 @@ import {
 import { Spinner } from "@/components/ui/spinner";
 import { useRevokeGrant } from "@/lib/mutations/credentials";
 import { useFetchGrants } from "@/lib/queries/credentials";
-import { timelockToMs } from "@/lib/time";
+import { timelockToMs, timelockToDate } from "@/lib/time";
 
 interface GrantsCenterProps {
   credentialId: string;
@@ -22,23 +22,6 @@ interface GrantsCenterProps {
 function generateGrantId(grant: idOSGrant): string {
   const { data_id, ag_grantee_wallet_identifier, locked_until } = grant;
   return [data_id, ag_grantee_wallet_identifier, locked_until].join("-");
-}
-
-// const MAX_REPRESENTABLE_TIMELOCK = 8.64e12;
-
-function timelockToDate(timelock: number): string {
-  if (!timelock) return "No timelock";
-
-  const milliseconds = timelockToMs(timelock);
-  const date = new Date(milliseconds);
-
-  if (Number.isNaN(date.getTime())) return timelock.toString();
-
-  return new Intl.DateTimeFormat(["ban", "id"], {
-    dateStyle: "short",
-    timeStyle: "short",
-    hour12: true,
-  }).format(date);
 }
 
 function Shares({ credentialId, grants }: { credentialId: string; grants: idOSGrant[] }) {
